Move event stream-related proto types and Java classes under `server.stream` package.
