+++ core-java/src/main/proto/spine/server/stream/event_record_filter.proto
@@ -19,30 +19,31 @@
 //
 syntax = "proto3";
 
-package spine.server;
+package spine.base;
 
 option java_generate_equals_and_hash = true;
 option java_multiple_files = true;
-option java_outer_classname = "EventStreamQueryProto";
-option java_package = "org.spine3.server";
+option java_outer_classname = "EventRecordFilterProto";
+option java_package = "org.spine3.server.stream";
 
-import public "google/protobuf/timestamp.proto";
-import public "spine/base/event_record_filter.proto";
+import public "spine/base/event.proto";
+//import public "spine/base/field_filter.proto";
 
-// The query specifies event to return from the event store.
-message EventStreamQuery {
-    // Include events with timestamp after this value.
-    //
-    // If this parameter is not specified, all events since the start of
-    // the history will be included.
-    google.protobuf.Timestamp after = 1;
+// Filters `EventRecord`s by the type of the event, event field values, and by `EventContext` field values.
+//
+// The time of the event is filtered by `before` and `after` attributes of `EventStreamQuery`.
+message EventRecordFilter {
+    // The full Proto type name of the event.
+    string event_type = 1;
 
-    // Include events with the timestamp before this value.
-    //
-    // If this parameter is not specified, all events till the end of
-    // the history will be included.
-    google.protobuf.Timestamp before = 2;
+    // The list of aggregate IDs of interest. Or empty, if include events for all aggregates.
+    repeated google.protobuf.Any aggregate_id = 2;
 
-    // Specification of `EventRecord`s to be included into resulting stream.
-    repeated spine.base.EventRecordFilter filter = 3;
+    //TODO:2016-01-11:alexander.yevsyukov: implement these filters if we need to query on deeper details.
+    // Accepted event field values.
+    // repeated FieldFilter event_field_filter = 3;
+
+    // Accepted `EventContext` field values.
+    // repeated FieldFilter context_field_filter = 4;
 }
+
