+++ core-java/src/test/java/org/spine3/server/stream/EventStoreServiceBuilderShould.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.server;
+package org.spine3.server.stream;
 
 import com.google.common.util.concurrent.MoreExecutors;
 import org.junit.Before;
