+++ examples/src/main/java/org/spine3/examples/eventstore/EventReader.java
@@ -28,8 +28,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.spine3.base.EventContext;
 import org.spine3.base.EventRecord;
-import org.spine3.server.EventStreamQuery;
-import org.spine3.server.grpc.EventStoreGrpc;
+import org.spine3.server.stream.EventStreamQuery;
+import org.spine3.server.stream.grpc.EventStoreGrpc;
 import org.spine3.type.TypeName;
 import org.spine3.util.EventRecords;
